rec-svc/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.3
pillow==10.3.0

torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu

faiss-cpu==1.8.0
transformers==4.42.3
tqdm==4.66.4
requests==2.32.3
